nvim: remove auto-completion suggestion from virtual text
So, having the fist auto-completion suggestion appear in the virtual
is nice for single line suggestions, but it does not play nicely
with the auto-completion pop-up window and multi line suggestion.
That is, the virtual text inserts multi lines of suggested text,
but these are covered by the auto-completion pop-up window. So, the
overall effect is the actual text being shunted down. The simplest
fix is to disable auto-completion suggestions in the virtual text and
rely only on the pop-up window.